+++ drools-framework-runtime-entity/src/main/java/org/chtijbug/drools/entity/history/process/BeforeProcessStartHistoryEvent.java
@@ -28,7 +28,7 @@ public class BeforeProcessStartHistoryEvent extends ProcessHistoryEvent {
     public BeforeProcessStartHistoryEvent() {
     }
 
-    public BeforeProcessStartHistoryEvent(int eventID, DroolsProcessInstanceObject processInstance, int ruleBaseId, int sessionId) {
+    public BeforeProcessStartHistoryEvent(Long eventID, DroolsProcessInstanceObject processInstance, int ruleBaseId, int sessionId) {
         super(eventID, ruleBaseId, sessionId);
         this.processInstance = processInstance;
     }
