+++ drools-framework-runtime-entity/src/main/java/org/chtijbug/drools/entity/history/session/SessionCreatedEvent.java
@@ -22,7 +22,7 @@ package org.chtijbug.drools.entity.history.session;
  * To change this template use File | Settings | File Templates.
  */
 public class SessionCreatedEvent extends SessionEvent {
-    public SessionCreatedEvent(int eventID, int ruleBaseId, int sessionId) {
+    public SessionCreatedEvent(Long eventID, int ruleBaseId, int sessionId) {
         super(eventID, ruleBaseId, sessionId);
     }
